Moved Actions to Catalog Methods
Tried using reflections for this, probably could figure it out later. For now, there's just a switch statement.

Removed the Actions folder entirely, moved it to Catalog methods, as you only ever run these actions on the catalog itself, so might as well be methods.

Had to fix up a bunch of tests for this as well, in addition to some minor changes to the Commandline app.